+++ core/src/main/kotlin/org/pastalab/fray/core/observers/AntithesisErrorReporter.kt
@@ -4,15 +4,9 @@ import com.antithesis.sdk.Assert
 import org.pastalab.fray.rmi.TestStatusObserver
 
 class AntithesisErrorReporter : TestStatusObserver {
-  override fun onExecutionStart() {}
-
-  override fun onExecutionDone(bugFound: Throwable?) {
-    if (bugFound != null) {
-      Assert.unreachable(
-          "An error occurred during the test execution: ${bugFound.message}, stack trace: ${bugFound.stackTraceToString()}",
-          null)
-    }
+  override fun onReportError(throwable: Throwable) {
+    Assert.unreachable(
+        "An error occurred during the test execution: ${throwable.message}, stack trace: ${throwable.stackTraceToString()}",
+        null)
   }
-
-  override fun saveToReportFolder(path: String) {}
 }